Move thumbnails app out of apps directory.
