+++ mapstory/settings/base.py
@@ -113,7 +113,7 @@ INSTALLED_APPS += (
     'solo',
     'coverage',
     'django_classification_banner',
-    'mapstory.apps.thumbnails',
+    'mapstory.thumbnails',
     'mapstory.journal',
     'mapstory.favorite',
     'mapstory.teams',
@@ -753,7 +753,7 @@ SCHEMA_DOWNLOAD_EXCLUDE = [
 ]
 
 # Choose thumbnail generator -- this is the delayed phantomjs generator
-THUMBNAIL_GENERATOR = "mapstory.apps.thumbnails.tasks.create_gs_thumbnail_mapstory_tx_aware"
+THUMBNAIL_GENERATOR = "mapstory.thumbnails.tasks.create_gs_thumbnail_mapstory_tx_aware"
 
 #
 # Classification banner
