+++ mapstory/thumbnails/static/thumbnails/static/thumbnail_storylayer.html
