+++ mapstory/thumbnails/tasks.py
