+++ mapstory/thumbnails/test_imgs/railroads-thumb-correct.png
