+++ mapstory/thumbnails/test_imgs/railroads-thumb-no_basemap.png
