+++ mapstory/thumbnails/thumbnail_storylayer.js
